+++ server/.env.example
@@ -1,2 +1,2 @@
 DB_URL=
-REDIRECT_URL=http://localhost:4000
+REDIRECT_URL=http://localhost:8080
